Realmテスト用のViewControllerを追加、Realmのエラーを修正
テスト用のViewControllerを作成、整理
汎用的に使用出来るUIViewクラス(UIViewUtil)を追加
Realmの実行時エラーを修正（Realmのプロパティに配列をもたせることはできない。プロパティの型にUInt32を使用できない)

+++ TangoBook/Extension/UIColor+ex.swift
@@ -10,7 +10,7 @@ import Foundation
 import UIKit
 
 extension UIColor {
-    func intColor() -> UInt32 {
+    func intColor() -> Int {
         var R : CGFloat = 0.0
         var G : CGFloat = 0.0
         var B : CGFloat = 0.0
@@ -18,9 +18,9 @@ extension UIColor {
         
         self.getRed(&R, green: &G, blue: &B, alpha: &A)
         
-        return (UInt32(A * 255.0) << 24) |
+        return Int((UInt32(A * 255.0) << 24) |
             (UInt32(R * 255.0) << 16) |
             (UInt32(G * 255.0) << 8) |
-            UInt32(B * 255.0)
+            UInt32(B * 255.0))
     }
 }
